+++ examples/benchmark_error.py
@@ -1,7 +1,7 @@
 from oersted.testing import make_helmholtz
 
 import oersted
-from oersted import CentroidMesh
+from oersted import CentroidMesh, OctreeSolver
 import numpy as np
 import matplotlib.pyplot as plt
 from matplotlib.ticker import ScalarFormatter
@@ -13,7 +13,7 @@ def main(nthetas: int = 1, size: float = 15.0, theta_min: float = 0.5, theta_max
     centroids, vol, jdensity = make_helmholtz(size)
     bdirect = oersted.b_field(CentroidMesh(centroids, vol), jdensity, centroids)
     for i, theta in enumerate(theta_vals):
-        boctree = oersted.bfield_octree(centroids, vol, jdensity, centroids, theta=theta, leaf_threshold=1)
+        boctree = oersted.b_field(CentroidMesh(centroids, vol), jdensity, centroids, solver=OctreeSolver(theta=theta))
         bmag_direct = np.linalg.norm(bdirect, axis=1)
         bmag_octree = np.linalg.norm(boctree, axis=1)
         # errs[i] = oersted.mean_relative_error(bmag_direct, bmag_octree)