+++ examples/example.py
@@ -1,14 +1,15 @@
-import oersted
+from oersted.testing import make_helmholtz
+from oersted import CentroidMesh, b_field, OctreeSolver
 from time import perf_counter
 
 size = 15.0
 theta = 0.5
 nthreads = 0
-centroids, vol, jdensity = oersted.testing.make_helmholtz(size)
+centroids, vol, jdensity = make_helmholtz(size)
+mesh = CentroidMesh(centroids, vol)
 
-# centroids = np.vstack((centroids, centroids))
-# vol = np.hstack((vol, vol))
-# jdensity = np.vstack((jdensity, jdensity))
+solver = OctreeSolver(theta=theta, n_threads=nthreads)
+
 
 print("Oersted Example - Helmholtz Problem")
 n = centroids.shape[0]
@@ -16,7 +17,7 @@ print(f"n = {n:.3e} ({n * n:.3e} interactions)")
 
 
 start = perf_counter()
-b = oersted.bfield_octree(centroids, vol, jdensity, centroids, theta=theta, nthreads=nthreads)
+b = b_field(mesh, jdensity, centroids, solver=solver)
 end = perf_counter()
 elapsed = end - start