+++ src/python.rs
@@ -12,7 +12,9 @@ use pyo3::prelude::*;
 use crate::biotsavart;
 #[cfg(feature = "parallel")]
 use crate::biotsavart_parallel;
+
 use crate::mesh;
+use crate::octree::{CurrentSources, HFieldSolver, Octree, point};
 use crate::vec3::Vec3;
 
 // ---
@@ -64,7 +66,6 @@ fn b_current_point_direct<'py>(
     tgt_pts: PyReadonlyArray2<f64>,
     nthreads_requested: u32,
 ) -> PyResult<Bound<'py, PyArray2<f64>>> {
-    
     // Transpose input data and allocate output arrays
     let n = tgt_pts.shape()[0];
     let (centx, centy, centz): (Vec<f64>, Vec<f64>, Vec<f64>) = pyarray_to_cols(src_pts);
@@ -96,89 +97,40 @@ fn b_current_point_direct<'py>(
 }
 
 #[pyfunction]
-fn _bfield_octree(
-    centx: PyReadonlyArray1<f64>,
-    centy: PyReadonlyArray1<f64>,
-    centz: PyReadonlyArray1<f64>,
-    vol: PyReadonlyArray1<f64>,
-    jx: PyReadonlyArray1<f64>,
-    jy: PyReadonlyArray1<f64>,
-    jz: PyReadonlyArray1<f64>,
-    x: PyReadonlyArray1<f64>,
-    y: PyReadonlyArray1<f64>,
-    z: PyReadonlyArray1<f64>,
-    mut bx: PyReadwriteArray1<f64>,
-    mut by: PyReadwriteArray1<f64>,
-    mut bz: PyReadwriteArray1<f64>,
+fn b_current_point_octree<'py>(
+    py: Python<'py>,
+    src_pts: PyReadonlyArray2<f64>,
+    src_vol: PyReadonlyArray1<f64>,
+    src_jdensity: PyReadonlyArray2<f64>,
+    tgt_pts: PyReadonlyArray2<f64>,
     theta: f64,
     leaf_threshold: u32,
     nthreads_requested: u32,
-) -> PyResult<()> {
-    // #[cfg(feature = "parallel")]
-    // if nthreads_requested != 1 {
-    //     use crate::biotsavart_parallel;
-    //     biotsavart_parallel::bfield_octree_parallel(
-    //         centx.as_slice()?,
-    //         centy.as_slice()?,
-    //         centz.as_slice()?,
-    //         vol.as_slice()?,
-    //         jx.as_slice()?,
-    //         jy.as_slice()?,
-    //         jz.as_slice()?,
-    //         x.as_slice()?,
-    //         y.as_slice()?,
-    //         z.as_slice()?,
-    //         bx.as_slice_mut()?,
-    //         by.as_slice_mut()?,
-    //         bz.as_slice_mut()?,
-    //         theta, leaf_threshold,
-    //         nthreads_requested
-    //     );
-    //     return Ok(());
-    // }
+) -> PyResult<Bound<'py, PyArray2<f64>>> {
+    // Transpose input data and allocate output arrays
+    let n = tgt_pts.shape()[0];
+    let (centx, centy, centz): (Vec<f64>, Vec<f64>, Vec<f64>) = pyarray_to_cols(src_pts);
+    let _src_vol: &[f64] = src_vol.as_slice()?;
+    let (jx, jy, jz): (Vec<f64>, Vec<f64>, Vec<f64>) = pyarray_to_cols(src_jdensity);
+    let (x, y, z): (Vec<f64>, Vec<f64>, Vec<f64>) = pyarray_to_cols(tgt_pts);
+    let (mut bx, mut by, mut bz): (Vec<f64>, Vec<f64>, Vec<f64>) = col_buffer(n);
 
-    use crate::octree::{CurrentSources, HFieldSolver, Octree, point};
-    let mut sources: CurrentSources<point::PointSources> =
-        CurrentSources(point::PointSources::new(
-            centx.as_slice()?,
-            centy.as_slice()?,
-            centz.as_slice()?,
-            vol.as_slice()?,
-            jx.as_slice()?,
-            jy.as_slice()?,
-            jz.as_slice()?,
-        ));
+    let mut sources: CurrentSources<point::PointSources> = CurrentSources(
+        point::PointSources::new((&centx, &centy, &centz), &_src_vol, (&jx, &jy, &jz)),
+    );
     let max_depth: u8 = 21;
     let tree: Octree<CurrentSources<point::PointSources>> =
         Octree::build_from_sources(sources, max_depth, leaf_threshold);
 
     #[cfg(feature = "parallel")]
     tree.h_field_parallel(
-        (x.as_slice()?, y.as_slice()?, z.as_slice()?),
-        (bx.as_slice_mut()?, by.as_slice_mut()?, bz.as_slice_mut()?),
+        (&x, &y, &z),
+        (&mut bx, &mut by, &mut bz),
         theta,
         nthreads_requested,
     );
 
-    // original version:
-    //
-    // biotsavart::bfield_octree(
-    //     centx.as_slice()?,
-    //     centy.as_slice()?,
-    //     centz.as_slice()?,
-    //     vol.as_slice()?,
-    //     jx.as_slice()?,
-    //     jy.as_slice()?,
-    //     jz.as_slice()?,
-    //     x.as_slice()?,
-    //     y.as_slice()?,
-    //     z.as_slice()?,
-    //     bx.as_slice_mut()?,
-    //     by.as_slice_mut()?,
-    //     bz.as_slice_mut()?,
-    //     theta, leaf_threshold
-    // );
-    Ok(())
+    Ok(cols_to_pyarray(py, (bx, by, bz)))
 }
 
 #[pyfunction]
@@ -611,7 +563,7 @@ fn _mesh_surface_tets<'py>(
 #[pymodule]
 fn _oersted<'py>(_py: Python, m: Bound<'py, PyModule>) -> PyResult<()> {
     m.add_function(wrap_pyfunction!(b_current_point_direct, m.clone())?)?;
-    m.add_function(wrap_pyfunction!(_bfield_octree, m.clone())?)?;
+    m.add_function(wrap_pyfunction!(b_current_point_octree, m.clone())?)?;
     m.add_function(wrap_pyfunction!(_bfield_dualtree, m.clone())?)?;
     m.add_function(wrap_pyfunction!(_bfield_hexahedron, m.clone())?)?;
     m.add_function(wrap_pyfunction!(_hfield_tetrahedrons, m.clone())?)?;