+++ tests/test_solenoid.py
@@ -13,7 +13,7 @@ Note: there's some sort of units mismatch with gmsh
 import numpy as np
 import matplotlib.pyplot as plt
 import oersted
-from oersted import Mesh, DirectSolver, CentroidMesh
+from oersted import DirectSolver, CentroidMesh, OctreeSolver
 from time import perf_counter
 
 #
@@ -23,7 +23,7 @@ from time import perf_counter
 datafile: str = "solenoid"
 remesh: bool = True
 theta: float = 0.1
-mesh_size: float = 15 # ~10M interactions; set to 33 for 1e6 interactions
+mesh_size: float = 15  # ~10M interactions; set to 33 for 1e6 interactions
 ntargets_axis: int = 25  # Along the axis
 nthreads: int = 0
 
@@ -31,8 +31,8 @@ nthreads: int = 0
 # Generate a mesh from a STEP file
 #
 
-mesh: Mesh = oersted.mesh.mesh_step(f"tests/data/{datafile}.stp", f"tests/data/{datafile}_mesh.csv", mesh_size, mesh_size)
-
+mesh = oersted.mesh.mesh_step(f"tests/data/{datafile}.stp", f"tests/data/{datafile}_mesh.csv", mesh_size, mesh_size)
+mesh = CentroidMesh(mesh.centroids, mesh.volumes)
 
 #
 # Setup sources from the mesh
@@ -56,7 +56,7 @@ targets_axis = np.zeros((ntargets_axis, 3))
 targets_axis[:, 2] = np.linspace(-0.125, 0.125, ntargets_axis)
 
 bdirect_axis = oersted.b_field(mesh, jdensity, targets_axis)
-boctree_axis = oersted.bfield_octree(mesh.centroids, mesh.volumes, jdensity, targets_axis, nthreads=nthreads, theta=theta)
+boctree_axis = oersted.b_field(mesh, jdensity, targets_axis, solver=OctreeSolver(n_threads=nthreads, theta=theta))
 
 #
 # Solve for self-fields
@@ -66,12 +66,12 @@ targets = mesh.centroids
 ntargets = targets.shape[0]
 
 start = perf_counter()
-bdirect = oersted.b_field(CentroidMesh(mesh.centroids,mesh.volumes), jdensity, targets, solver=DirectSolver(n_threads=nthreads))
+bdirect = oersted.b_field(CentroidMesh(mesh.centroids, mesh.volumes), jdensity, targets, solver=DirectSolver(n_threads=nthreads))
 end = perf_counter()
 direct_elapsed = end - start
 
 start = perf_counter()
-boctree = oersted.bfield_octree(mesh.centroids, mesh.volumes, jdensity, targets, nthreads=nthreads, theta=theta)
+boctree = oersted.b_field(mesh, jdensity, targets, solver=OctreeSolver(n_threads=nthreads, theta=theta))
 end = perf_counter()
 octree_elapsed = end - start
 
@@ -115,10 +115,10 @@ plt.savefig("tests/fig/solenoid_test.svg")
 
 
 def test_solenoid():
-    print(f"Error in mesh, direct vs octree: {err_mesh*100:.2f} %")
+    print(f"Error in mesh, direct vs octree: {err_mesh * 100:.2f} %")
     assert err_mesh < 1e-2
     assert err_axis < 1e-2
 
 
-if __name__=="__main__":
+if __name__ == "__main__":
     test_solenoid()